fix typos and paging (#149)
* fix typos and paging

* Update

* update

* update

* update

* Update

+++ aivalut/src/ml/metrics.h
@@ -9,8 +9,8 @@
 #include <vector>
 using namespace std;
 
-inline float trapz(vector<float> x, vector<float> y) {
-  float res = 0;
+inline double trapz(vector<double> x, vector<double> y) {
+  double res = 0;
   int num_elements = x.size();
   for (int i = 1; i < num_elements; i++) {
     res += (x[i] - x[i - 1]) * (y[i] + y[i - 1]) / 2;
@@ -18,9 +18,9 @@ inline float trapz(vector<float> x, vector<float> y) {
   return res;
 }
 
-inline vector<float> get_thresholds_idxs(vector<float> y_pred) {
-  vector<float> thresholds_idxs;
-  set<float> s{};
+inline vector<double> get_thresholds_idxs(vector<double> y_pred) {
+  vector<double> thresholds_idxs;
+  set<double> s{};
   reverse(y_pred.begin(), y_pred.end());
   int y_pred_size = y_pred.size();
   for (int i = 0; i < y_pred_size; i++) {
@@ -31,7 +31,7 @@ inline vector<float> get_thresholds_idxs(vector<float> y_pred) {
   return thresholds_idxs;
 }
 
-inline float roc_auc_score(vector<float> y_pred, vector<int> y_true) {
+inline double roc_auc_score(vector<double> y_pred, vector<int> y_true) {
   int num_elements = y_pred.size();
   vector<int> temp_idxs(num_elements);
   iota(temp_idxs.begin(), temp_idxs.end(), 0);
@@ -44,9 +44,9 @@ inline float roc_auc_score(vector<float> y_pred, vector<int> y_true) {
   }
   sort(y_pred.begin(), y_pred.end());
 
-  vector<float> thresholds_idxs = get_thresholds_idxs(y_pred);
+  vector<double> thresholds_idxs = get_thresholds_idxs(y_pred);
 
-  vector<float> tps = {0};
+  vector<double> tps = {0};
   for (int i = 1; i < num_elements; i++) {
     tps.push_back(y_true[i] + tps[i - 1]);
   }
@@ -54,7 +54,7 @@ inline float roc_auc_score(vector<float> y_pred, vector<int> y_true) {
     tps[i] = tps[i] / tps[tps.size() - 1];
   }
 
-  vector<float> fps = {1};
+  vector<double> fps = {1};
   for (int i = 1; i < num_elements; i++) {
     fps.push_back(1 - y_true[i] + fps[i - 1]);
   }
@@ -65,23 +65,23 @@ inline float roc_auc_score(vector<float> y_pred, vector<int> y_true) {
   return trapz(tps, fps);
 }
 
-inline float ovr_roc_auc_score(vector<vector<float>> y_pred,
-                               vector<int> y_true) {
+inline double ovr_roc_auc_score(vector<vector<double>> y_pred,
+                                vector<int> y_true) {
   int num_elements = y_pred.size();
   int num_classes = y_pred[0].size();
 
   if (num_classes == 2) {
-    vector<float> y_pred_pos(num_elements, 0);
+    vector<double> y_pred_pos(num_elements, 0);
     for (int i = 0; i < num_elements; i++) {
       y_pred_pos[i] = y_pred[i][1];
     }
     return roc_auc_score(y_pred_pos, y_true);
   } else {
-    float ovr_average_score = 0;
-    float tmp_roc_auc_score = 0;
-    float count_classes = 0;
+    double ovr_average_score = 0;
+    double tmp_roc_auc_score = 0;
+    double count_classes = 0;
     for (int c = 0; c < num_classes; c++) {
-      vector<float> y_pred_c(num_elements, 0);
+      vector<double> y_pred_c(num_elements, 0);
       vector<int> y_true_c(num_elements, 0);
       for (int i = 0; i < num_elements; i++) {
         y_pred_c[i] = y_pred[i][c];
@@ -100,11 +100,11 @@ inline float ovr_roc_auc_score(vector<vector<float>> y_pred,
   }
 }
 
-inline float ovr_roc_auc_score(vector<vector<float>> y_pred,
-                               vector<float> y_true) {
+inline double ovr_roc_auc_score(vector<vector<double>> y_pred,
+                                vector<double> y_true) {
   std::vector<int> out;
   out.reserve(y_pred.size());
   std::transform(y_true.begin(), y_true.end(), std::back_inserter(out),
-                 [](float n) { return (int)(n); });
+                 [](double n) { return (int)(n); });
   return ovr_roc_auc_score(y_pred, out);
 }
